Module wasmer_singlepass_backend::codegen_x64

Structs

CONSTRUCT_STACK_AND_CALL_WASM
Performs a System V call to target with [stack_top..stack_base] as the argument list, from right to left.

CacheImage
On-disk cache format. Offsets are relative to the start of the executable image.

CodegenConfig

CodegenError

ControlFrame

FuncPtr

SinglepassCache

X64ModuleCodeGenerator

X64FunctionCode

X64ExecutionContext

Enums

FuncPtrInner

IfElseState

Constants

GEF32_LT_I32_MIN
Greatest Exact Float (32 bits) less-than i32::MIN when rounding towards zero.

GEF32_LT_I64_MIN
Greatest Exact Float (32 bits) less-than i64::MIN when rounding towards zero.

GEF32_LT_U32_MIN
Greatest Exact Float (32 bits) less-than u32::MIN when rounding towards zero.

GEF32_LT_U64_MIN
Greatest Exact Float (32 bits) less-than u64::MIN when rounding towards zero.

GEF64_LT_I32_MIN
Greatest Exact Float (64 bits) less-than i32::MIN when rounding towards zero.

GEF64_LT_I64_MIN
Greatest Exact Float (64 bits) less-than i64::MIN when rounding towards zero.

GEF64_LT_U32_MIN
Greatest Exact Float (64 bits) less-than u32::MIN when rounding towards zero.

GEF64_LT_U64_MIN
Greatest Exact Float (64 bits) less-than u64::MIN when rounding towards zero.

INLINE_BREAKPOINT_SIZE_AARCH64_SINGLEPASS
Inline breakpoint size for aarch64.

INLINE_BREAKPOINT_SIZE_X86_SINGLEPASS
Inline breakpoint size for x86-64.

LEF32_GT_I32_MAX
Least Exact Float (32 bits) greater-than i32::MAX when rounding towards zero.

LEF32_GT_I64_MAX
Least Exact Float (32 bits) greater-than i64::MAX when rounding towards zero.

LEF32_GT_U32_MAX
Least Exact Float (32 bits) greater-than u32::MAX when rounding towards zero.

LEF32_GT_U64_MAX
Least Exact Float (32 bits) greater-than u64::MAX when rounding towards zero.

LEF64_GT_I32_MAX
Least Exact Float (64 bits) greater-than i32::MAX when rounding towards zero.

LEF64_GT_I64_MAX
Least Exact Float (64 bits) greater-than i64::MAX when rounding towards zero.

LEF64_GT_U32_MAX
Least Exact Float (64 bits) greater-than u32::MAX when rounding towards zero.

LEF64_GT_U64_MAX
Least Exact Float (64 bits) greater-than u64::MAX when rounding towards zero.

_IMPL_DESERIALIZE_FOR_CacheImage

_IMPL_SERIALIZE_FOR_CacheImage

Statics

ARCH

BACKEND_ID

CONSTRUCT_STACK_AND_CALL_WASM

Functions

get_location_released

sort_call_movs

type_to_wp_type