Struct wasmer_singlepass_backend::FunctionCodeGenerator

pub struct FunctionCodeGenerator {
    local_function_id: usize,
    signatures: Arc<Map<SigIndex, FuncSig>>,
    function_signatures: Arc<Map<FuncIndex, SigIndex>>,
    fsm: FunctionStateMap,
    offset: usize,
    assembler: Option<Assembler>,
    function_labels: Option<HashMap<usize, (DynamicLabel, Option<AssemblyOffset>)>>,
    breakpoints: Option<HashMap<AssemblyOffset, Box<dyn Fn(BreakpointInfo) -> Result<(), Box<dyn Any + Send>> + Send + Sync + 'static>>>,
    returns: SmallVec<[WpType; 1]>,
    locals: Vec<Location>,
    num_params: usize,
    num_locals: usize,
    value_stack: Vec<Location>,
    control_stack: Vec<ControlFrame>,
    machine: Machine,
    unreachable_depth: usize,
    config: Arc<CodegenConfig>,
    exception_table: Option<ExceptionTable>,
}

Fields
local_function_id: usizesignatures: Arc<Map<SigIndex, FuncSig>>function_signatures: Arc<Map<FuncIndex, SigIndex>>fsm: FunctionStateMapoffset: usizeassembler: Option<Assembler>function_labels: Option<HashMap<usize, (DynamicLabel, Option<AssemblyOffset>)>>breakpoints: Option<HashMap<AssemblyOffset, Box<dyn Fn(BreakpointInfo) -> Result<(), Box<dyn Any + Send>> + Send + Sync + 'static>>>returns: SmallVec<[WpType; 1]>locals: Vec<Location>num_params: usizenum_locals: usizevalue_stack: Vec<Location>control_stack: Vec<ControlFrame>machine: Machineunreachable_depth: usizeconfig: Arc<CodegenConfig>exception_table: Option<ExceptionTable>
Methods

impl X64FunctionCode

fn mark_trappable(
a: &mut Assembler,
m: &Machine,
fsm: &mut FunctionStateMap,
control_stack: &mut [ControlFrame]
)

fn mark_range_with_exception_code<F: FnOnce(&mut Assembler) -> R, R>(
a: &mut Assembler,
etable: &mut ExceptionTable,
code: ExceptionCode,
f: F
) -> R

Marks each address in the code range emitted by f with the exception code code.

fn emit_relaxed_xdiv(
a: &mut Assembler,
m: &mut Machine,
etable: &mut ExceptionTable,
op: fn(_: &mut Assembler, _: Size, _: Location),
sz: Size,
loc: Location,
fsm: &mut FunctionStateMap,
control_stack: &mut [ControlFrame]
)

Moves loc to a valid location for div/idiv.

fn emit_relaxed_zx_sx(
a: &mut Assembler,
m: &mut Machine,
op: fn(_: &mut Assembler, _: Size, _: Location, _: Size, _: Location),
sz_src: Size,
src: Location,
sz_dst: Size,
dst: Location
) -> Result<(), CodegenError>

Moves src and dst to valid locations for movzx/movsx.

fn emit_relaxed_binop(
a: &mut Assembler,
m: &mut Machine,
op: fn(_: &mut Assembler, _: Size, _: Location, _: Location),
sz: Size,
src: Location,
dst: Location
)

Moves src and dst to valid locations for generic instructions.

fn emit_relaxed_avx(
a: &mut Assembler,
m: &mut Machine,
op: fn(_: &mut Assembler, _: XMM, _: XMMOrMemory, _: XMM),
src1: Location,
src2: Location,
dst: Location
) -> Result<(), CodegenError>

Moves src1 and src2 to valid locations and possibly adds a layer of indirection for dst for AVX instructions.

fn emit_relaxed_avx_base<F: FnOnce(&mut Assembler, &mut Machine, XMM, XMMOrMemory, XMM)>(
a: &mut Assembler,
m: &mut Machine,
op: F,
src1: Location,
src2: Location,
dst: Location
) -> Result<(), CodegenError>

Moves src1 and src2 to valid locations and possibly adds a layer of indirection for dst for AVX instructions.

fn emit_binop_i32(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
f: fn(_: &mut Assembler, _: Size, _: Location, _: Location)
)

I32 binary operation with both operands popped from the virtual stack.

fn emit_binop_i64(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
f: fn(_: &mut Assembler, _: Size, _: Location, _: Location)
)

I64 binary operation with both operands popped from the virtual stack.

fn emit_cmpop_i32_dynamic_b(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
c: Condition,
loc_b: Location
) -> Result<(), CodegenError>

I32 comparison with loc_b from input.

fn emit_cmpop_i32(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
c: Condition
) -> Result<(), CodegenError>

I32 comparison with both operands popped from the virtual stack.

fn emit_cmpop_i64_dynamic_b(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
c: Condition,
loc_b: Location
) -> Result<(), CodegenError>

I64 comparison with loc_b from input.

fn emit_cmpop_i64(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
c: Condition
) -> Result<(), CodegenError>

I64 comparison with both operands popped from the virtual stack.

fn emit_xcnt_i32(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
f: fn(_: &mut Assembler, _: Size, _: Location, _: Location)
) -> Result<(), CodegenError>

I32 lzcnt/tzcnt/popcnt with operand popped from the virtual stack.

fn emit_xcnt_i64(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
f: fn(_: &mut Assembler, _: Size, _: Location, _: Location)
) -> Result<(), CodegenError>

I64 lzcnt/tzcnt/popcnt with operand popped from the virtual stack.

fn emit_shift_i32(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
f: fn(_: &mut Assembler, _: Size, _: Location, _: Location)
)

I32 shift with both operands popped from the virtual stack.

fn emit_shift_i64(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
f: fn(_: &mut Assembler, _: Size, _: Location, _: Location)
)

I64 shift with both operands popped from the virtual stack.

fn emit_fp_binop_avx(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
f: fn(_: &mut Assembler, _: XMM, _: XMMOrMemory, _: XMM)
) -> Result<(), CodegenError>

Floating point (AVX) binary operation with both operands popped from the virtual stack.

fn emit_fp_cmpop_avx(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
f: fn(_: &mut Assembler, _: XMM, _: XMMOrMemory, _: XMM)
) -> Result<(), CodegenError>

Floating point (AVX) comparison with both operands popped from the virtual stack.

fn emit_fp_unop_avx(
a: &mut Assembler,
m: &mut Machine,
value_stack: &mut Vec<Location>,
f: fn(_: &mut Assembler, _: XMM, _: XMMOrMemory, _: XMM)
) -> Result<(), CodegenError>

Floating point (AVX) binary operation with both operands popped from the virtual stack.

fn emit_call_sysv<I: Iterator<Item = Location>, F: FnOnce(&mut Assembler)>(
a: &mut Assembler,
m: &mut Machine,
cb: F,
params: I,
state_context: Option<(&mut FunctionStateMap, &mut [ControlFrame])>
) -> Result<(), CodegenError>

Emits a System V call sequence.

This function must not use RAX before cb is called.

fn emit_call_sysv_label<I: Iterator<Item = Location>>(
a: &mut Assembler,
m: &mut Machine,
label: DynamicLabel,
params: I,
state_context: Option<(&mut FunctionStateMap, &mut [ControlFrame])>
) -> Result<(), CodegenError>

Emits a System V call sequence, specialized for labels as the call target.

fn emit_memory_op<F: FnOnce(&mut Assembler, &mut Machine, GPR) -> Result<(), CodegenError>>(
module_info: &ModuleInfo,
config: &CodegenConfig,
a: &mut Assembler,
m: &mut Machine,
etable: &mut ExceptionTable,
addr: Location,
memarg: &MemoryImmediate,
check_alignment: bool,
value_size: usize,
cb: F
) -> Result<(), CodegenError>

Emits a memory operation.

fn emit_compare_and_swap<F: FnOnce(&mut Assembler, &mut Machine, GPR, GPR)>(
module_info: &ModuleInfo,
config: &CodegenConfig,
a: &mut Assembler,
m: &mut Machine,
etable: &mut ExceptionTable,
loc: Location,
target: Location,
ret: Location,
memarg: &MemoryImmediate,
value_size: usize,
memory_sz: Size,
stack_sz: Size,
cb: F
) -> Result<(), CodegenError>

Emits a memory operation.

fn emit_f32_int_conv_check(
a: &mut Assembler,
m: &mut Machine,
reg: XMM,
lower_bound: f32,
upper_bound: f32,
underflow_label: <Assembler as Emitter>::Label,
overflow_label: <Assembler as Emitter>::Label,
nan_label: <Assembler as Emitter>::Label,
succeed_label: <Assembler as Emitter>::Label
)

fn emit_f32_int_conv_check_trap(
a: &mut Assembler,
m: &mut Machine,
etable: &mut ExceptionTable,
reg: XMM,
lower_bound: f32,
upper_bound: f32
)

fn emit_f32_int_conv_check_sat<F1: FnOnce(&mut Assembler, &mut Machine), F2: FnOnce(&mut Assembler, &mut Machine), F3: FnOnce(&mut Assembler, &mut Machine), F4: FnOnce(&mut Assembler, &mut Machine)>(
a: &mut Assembler,
m: &mut Machine,
reg: XMM,
lower_bound: f32,
upper_bound: f32,
underflow_cb: F1,
overflow_cb: F2,
nan_cb: Option<F3>,
convert_cb: F4
)

fn emit_f64_int_conv_check(
a: &mut Assembler,
m: &mut Machine,
reg: XMM,
lower_bound: f64,
upper_bound: f64,
underflow_label: <Assembler as Emitter>::Label,
overflow_label: <Assembler as Emitter>::Label,
nan_label: <Assembler as Emitter>::Label,
succeed_label: <Assembler as Emitter>::Label
)

fn emit_f64_int_conv_check_trap(
a: &mut Assembler,
m: &mut Machine,
etable: &mut ExceptionTable,
reg: XMM,
lower_bound: f64,
upper_bound: f64
)

fn emit_f64_int_conv_check_sat<F1: FnOnce(&mut Assembler, &mut Machine), F2: FnOnce(&mut Assembler, &mut Machine), F3: FnOnce(&mut Assembler, &mut Machine), F4: FnOnce(&mut Assembler, &mut Machine)>(
a: &mut Assembler,
m: &mut Machine,
reg: XMM,
lower_bound: f64,
upper_bound: f64,
underflow_cb: F1,
overflow_cb: F2,
nan_cb: Option<F3>,
convert_cb: F4
)

pub fn get_state_diff(
m: &Machine,
fsm: &mut FunctionStateMap,
control_stack: &mut [ControlFrame]
) -> usize